Don't overflow trying to draw ticks on ruler (KICAD-9HV).

+++ common/eda_shape.cpp
@@ -1020,8 +1020,8 @@ int EDA_SHAPE::GetRadius() const
     }
 
     // don't allow degenerate circles/arcs
-    if( radius > INT_MAX / 2.0 )
-        radius = INT_MAX / 2.0;
+    if( radius > (double) INT_MAX / 2.0 )
+        radius = (double) INT_MAX / 2.0;
 
     return std::max( 1, KiROUND( radius ) );
 }
